# docs/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dubcast Radio API",
    "description": "API интернет-радио Dubcast.\nСодержит публичные, пользовательские и админские эндпоинты.\n",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8089",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/tracks/{id}": {
      "get": {
        "tags": [
          "Tracks (Admin)"
        ],
        "summary": "Get track by ID",
        "description": "Returns a single track by its identifier.\n",
        "operationId": "getById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Track ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 42
          }
        ],
        "responses": {
          "404": {
            "description": "Track not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Track found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Tracks (Admin)"
        ],
        "summary": "Update track",
        "description": "Updates an existing track by ID.\nAll updatable fields are taken from the request body.\n",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Track ID to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 42
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Track not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Track updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error in request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Tracks (Admin)"
        ],
        "summary": "Delete track",
        "description": "Deletes an existing track by its ID.\nTypically used to clean up invalid or obsolete tracks.\n",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Track ID to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 42
          }
        ],
        "responses": {
          "404": {
            "description": "Track not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "Track deleted (no content)"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/tracks": {
      "get": {
        "tags": [
          "Tracks (Admin)"
        ],
        "summary": "Get all tracks",
        "description": "Returns the full list of tracks in the catalog.\nCan be used in admin UI to browse or search tracks.\n",
        "operationId": "getAll_1",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "List of tracks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tracks (Admin)"
        ],
        "summary": "Create track",
        "description": "Creates a new track in the catalog.\nTypically used from admin tools or import flows.\n",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Track created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error in request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (no or invalid JWT)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (user is not an admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "validationErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "TrackDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "soundcloudUrl": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "artworkUrl": {
            "type": "string"
          }
        },
        "required": [
          "durationSeconds",
          "soundcloudUrl",
          "title"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}